Change parameter contract name to @NotNullIfReturns
